Reduce the copy of bytes when retrieving the encoded bytes (#5)
### Motivation

Currently the encoded bytes are retrieved from the `OutputStream` by the
`snapshot` function. However, it returns a `std::vector<uint8_t>` but we
need a `std::string`. So we have to make a copy again.

### Modifications

Create an `InputStream` from the `OutputStream` and use the `next`
method to read bytes into the pre-allocated `std::string` from it.